add metrics APIs (metrics.k8s.io)
Add APIs for `metrics.k8s.io/v1beta1`.

Kubernetes OpenAPI specs do not include metrics endpoints. This PR update the specifications first for that, and then uses it to generate the required code. Both the updated specifications and generated code is included in this PR. These changes need to be merged when we upgrade to a newer specification in future.