also allow retries of mutating APIs
State changing APIs are not retried by default. But added an option to switch that on when the caller desires.
Should be safe when retried only on 50x server errors or connect errors.

The option can be switched on for the KuberContext by calling the `set_retries` method.
Alternatively each API call also accepts a `max_retries` keywork argument to override the default set in the context.