allow loading of other k8s API implementations
Refactor to move all code generated from swagger spec into a separate module, named `ApiImpl`.
`KuberContext` now optionally takes a module reference to load APIs from. It defaults to the one bundled in Kuber.jl - `ApiImpl`, but can be replaced with a different module when `KuberContext` is constructed. This will let Kuber.jl be used with code generated for a different Kubernetes API version, or for a spec with custom APIs (e.g. for CRDs).